requirements-dev.txt
black==24.10.0
isort==5.13.2
ruff==0.8.2
mypy==1.11.2
pytest==8.3.3
types-requests==2.32.0.20240914
python-dotenv==1.1.1

